Move DataManager class to new __init__.py file
